Add files via upload
Datasets task - For this project, I was tasked with using select functions within Pandas in Python for a data analysis related task using datasets. 
Harnessing data from a raw datasets (balance.txt), I used the iloc function to select specific columns from the dataset to gain insights from the data. Beyond this, there are sorting exercises so as to show, for example, users with a high education value. Finally, with instructions to select specific rows, columns or specific data from the dataset, I use the iloc function.

Reports task - This task was centred around creating a report from a large, raw dataset csv that provides the following information:
- Compare the average income based on ethnicity.
- On average, do married or single people have a higher balance?
- What is the highest income in our dataset?
- What is the lowest income in our dataset?
- How many cards do we have recoreded in our dataset?
- How many females do we have information for vs how many males?

Again, using Jupyter and Python, we can see the use of the print function to present the results of these questions, using the mean function for average, for example, or the max function to find the highest income in the dataset. The sum function is then used to find the total number of cards recorded and, finally, count is used to determine the total number of females and males